+++ R/metabin.R
@@ -142,6 +142,8 @@
 #'   grouping variable should be printed in front of the group labels.
 #' @param byseparator A character string defining the separator
 #'   between label and levels of grouping variable.
+#' @param test.subgroup A logical value indicating whether to print
+#'   results of test for subgroup differences.
 #' @param print.CMH A logical indicating whether result of the
 #'   Cochran-Mantel-Haenszel test for overall effect should be
 #'   printed.
@@ -877,10 +879,12 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                     title = gs("title"), complab = gs("complab"),
                     outclab = "",
                     label.e = gs("label.e"), label.c = gs("label.c"),
-                    label.left = gs("label.left"), label.right = gs("label.right"),
+                    label.left = gs("label.left"),
+                    label.right = gs("label.right"),
                     ##
                     byvar, bylab, print.byvar = gs("print.byvar"),
                     byseparator = gs("byseparator"),
+                    test.subgroup = gs("test.subgroup"),
                     ##
                     print.CMH = gs("print.CMH"),
                     ##
@@ -1119,8 +1123,10 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   if (length(incr) > 1)
     chklength(incr, k.All, fun)
   ##
-  if (by)
+  if (by) {
     chklength(byvar, k.All, fun)
+    chklogical(test.subgroup)
+  }
   ##
   ## Additional checks
   ##
@@ -1967,6 +1973,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
     res$bylab <- bylab
     res$print.byvar <- print.byvar
     res$byseparator <- byseparator
+    res$test.subgroup <- test.subgroup
     res$tau.common <- tau.common
     ##
     if (!tau.common)