+++ R/metacont.R
@@ -153,7 +153,9 @@
 #'   grouping variable should be printed in front of the group labels.
 #' @param byseparator A character string defining the separator
 #'   between label and levels of grouping variable.
-#' @param keepdata A logical indicating whether original data (set)
+#' @param test.subgroup A logical value indicating whether to print
+#'   results of test for subgroup differences.
+##' @param keepdata A logical indicating whether original data (set)
 #'   should be kept in meta object.
 #' @param warn A logical indicating whether warnings should be printed
 #'   (e.g., if studies are excluded from meta-analysis due to zero
@@ -831,6 +833,7 @@ metacont <- function(n.e, mean.e, sd.e, n.c, mean.c, sd.c, studlab,
                      ##
                      byvar, bylab, print.byvar = gs("print.byvar"),
                      byseparator = gs("byseparator"),
+                     test.subgroup = gs("test.subgroup"),
                      ##
                      keepdata = gs("keepdata"),
                      warn = gs("warn"),
@@ -1145,8 +1148,10 @@ metacont <- function(n.e, mean.e, sd.e, n.c, mean.c, sd.c, studlab,
     approx.sd.c <- setchar(approx.sd.c, c("", "iqr.range", "iqr", "range"))
   }
   ##
-  if (by)
+  if (by) {
     chklength(byvar, k.All, arg)
+    chklogical(test.subgroup)
+  }
   ##
   ## Additional checks
   ##
@@ -1956,6 +1961,7 @@ metacont <- function(n.e, mean.e, sd.e, n.c, mean.c, sd.c, studlab,
     res$bylab <- bylab
     res$print.byvar <- print.byvar
     res$byseparator <- byseparator
+    res$test.subgroup <- test.subgroup
     res$tau.common <- tau.common
     ##
     if (!tau.common)