+++ R/metaprop.R
@@ -111,6 +111,8 @@
 #'   grouping variable should be printed in front of the group labels.
 #' @param byseparator A character string defining the separator
 #'   between label and levels of grouping variable.
+#' @param test.subgroup A logical value indicating whether to print
+#'   results of test for subgroup differences.
 #' @param keepdata A logical indicating whether original data (set)
 #'   should be kept in meta object.
 #' @param warn A logical indicating whether the addition of
@@ -869,6 +871,7 @@ metaprop <- function(event, n, studlab,
                      ##
                      byvar, bylab, print.byvar = gs("print.byvar"),
                      byseparator = gs("byseparator"),
+                     test.subgroup = gs("test.subgroup"),
                      ##
                      keepdata = gs("keepdata"),
                      warn = gs("warn"),
@@ -1025,8 +1028,10 @@ metaprop <- function(event, n, studlab,
   if (length(incr) > 1)
     chklength(incr, k.All, fun)
   ##
-  if (by)
+  if (by) {
     chklength(byvar, k.All, fun)
+    chklogical(test.subgroup)
+  }
   ##
   ## Additional checks
   ##
@@ -1602,6 +1607,7 @@ metaprop <- function(event, n, studlab,
     res$bylab <- bylab
     res$print.byvar <- print.byvar
     res$byseparator <- byseparator
+    res$test.subgroup <- test.subgroup
     res$tau.common <- tau.common
     ##
     if (!tau.common)